% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APES.R
\name{APES}
\alias{APES}
\title{Perform APES model selection}
\usage{
APES(model, k = NULL, estimator = "leaps", time_limit = 10L, verbose = FALSE)
}
\arguments{
\item{model}{A "glm" class object}

\item{k}{Model size to explore. Default to NULL, which searches through all variables in the model.
Alternatively, user can input a vector. If estimator is:
\itemize{
 \item "leaps", then models up to size max(k) wil be explored.
 \item "mio", then models with specified values will be explored.
}}

\item{estimator}{Either "leaps" (default) or "mio", which correspond to
optimisation algorithms available in
the leaps and bestsubset package, respectively.}

\item{time_limit}{The time limit for the maximum time allocated to each
model size model when the "mio" estimator was selected. It will not affect the speed if leaps.}

\item{verbose}{Whether to print off messages duringcomputations}
}
\description{
Perform APES model selection
}
\examples{
set.seed(10)
n = 100
p = 10
k = 1:10
beta = c(1, -1, rep(0, p-2))
x = matrix(rnorm(n*p), ncol = p)
colnames(x) = paste0("X", 1:p)
y = rbinom(n = n, size = 1, prob = expit(x \%*\% beta))
data = data.frame(y, x)
model = glm(y ~ ., data = data, family = "binomial")
APES(model = model)

y = rpois(n = n, lambda = exp(x \%*\% beta))
data = data.frame(y, x)
model = glm(y ~ ., data = data, family = "poisson")
APES(model = model)
}
