% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apes_cox.R
\name{apes_cox}
\alias{apes_cox}
\title{Perform APES Cox PH regression using either the leaps or bestsubset package}
\usage{
apes_cox(x, time, status, k, estimator = "leaps", time.limit = 60)
}
\arguments{
\item{x}{The n by p design matrix for logistic regression}

\item{time}{survival time in survival analysis}

\item{status}{survival status in survival analysis}

\item{k}{Model size to explore. If "leaps" was selected as the estimator, then model up to size max(k) will be explored.
If "mio" was selected as the estimator, then model sizes specified in k will be explored.}

\item{estimator}{Either "leaps" or "mio", which correspond to optimisation algorithms available in the leaps and bestsubset package, respectively.}

\item{time.limit}{The time limit for the maximum time allocated to each model size model when the "mio" estimator was selected. It will not affect the speed if leaps}
}
\value{
A list of results:
\itemize{
\item{apesModelDf:} A data.frame (tibble) containing the optimal model of each size and associated statistics
\item{apesMleBeta:} A matrix of MLE for the most optimal model of each model size.
\item{apesMleBetaBinary:} A data.frame (tibble), which has a column indicating if
a variable has been selected into the most optimal model of each size.
It is a binarised, linear version of the previous \code{apesMleBeta} output.
\item{apesTimeDiff:} A difftime object of the computational time.
\item{selectedModelBeta:} A matrix containing the MLE correspond to the AIC and BIC selected model. A subset of the \code{apesMleBeta} output.
\item{aicWeights:} A matrix of weights corresponding to the AIC_j/sum(AIC_j), where AIC_j denotes the most optimal model of the size j in the \code{apesModelDf} output.
\item{bicWeights:} A matrix of weights corresponding to the BIC_j/sum(BIC_j), where BIC_j denotes the most optimal model of the size j in the \code{apesModelDf} output.
\item{modelAvgBeta}: A matrix of AIC and BIC weighted coefficients. It is identical to the matrix product of apesMleBeta and aicWeights (or bicWeights).
\item{responseTibble}: A data.frame (tibble) of estimates for each observations, e.g. initial fitted response, final fitted response.
}
}
\description{
Perform APES Cox PH regression using either the leaps or bestsubset package
}
\examples{
set.seed(10101)
n = 200
p = 10
nzc = p/3
x = matrix(rnorm(n*p), n, p)
colnames(x) = paste0("X", 1:p)
beta = runif(round(nzc))
fx = x[,seq(nzc)] \%*\% beta
hx = exp(fx)
ty = rexp(n, hx)
tcens = rbinom(n = n, prob = 0.3, size = 1) # censoring indicator
time = ty
status = 1-tcens
apesLeapsResult = apes_cox(x = x, time = time, status = status, k = p,
estimator = "leaps")
\dontrun{
apesMioResult = apes_cox(x = x, time = time, status = status, k = 1:p,
estimator = "mio")
}
}
