% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apes_leaps.R
\name{apes_leaps}
\alias{apes_leaps}
\title{Perform APES using the leaps package}
\usage{
apes_leaps(x, y, Pi, maxK)
}
\arguments{
\item{x}{n times p design matrix for logistics regression}

\item{y}{response vector for logistics regression of length n}

\item{Pi}{esimated probability of each observation using a baseline model. Of length n.}

\item{maxK}{maximum model size, less than p.}
}
\description{
Perform APES using the leaps package
}
\examples{
set.seed(10)
n = 100
p = 10
maxK = 5
x = matrix(rnorm(n*p), ncol = p)
colnames(x) = paste0("X", 1:p)
y = round(runif(n, 0,1))
Pi = runif(n, 0, 1)
apes_leaps(x = x, y = y, Pi = Pi, maxK = maxK)
}
