% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apes_leaps_logit.R
\name{apes_leaps_logit}
\alias{apes_leaps_logit}
\title{Perform APES-logistics regression using the leaps package}
\usage{
apes_leaps_logit(x, y, Pi, maxK)
}
\arguments{
\item{x}{The n by p design matrix for logistics regression}

\item{y}{The response vector for logistics regression of length n}

\item{Pi}{Esimated probability of each observation using a baseline model. It should be a vector of length n.}

\item{maxK}{The maximum model size to explore, this value must be less than p.}
}
\description{
Perform APES-logistics regression using the leaps package
}
\examples{

set.seed(10)
n = 100
p = 10
maxK = 5
x = matrix(rnorm(n*p), ncol = p)
colnames(x) = paste0("X", 1:p)
y = round(runif(n, 0,1))
Pi = runif(n, 0, 1)
apesResult = apes_leaps_logit(x = x, y = y, Pi = Pi, maxK = maxK)
names(apesResult)
}
