% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apes_leap_poisson.R
\name{apes_leaps_poisson}
\alias{apes_leaps_poisson}
\title{Perform APES using the leaps package}
\usage{
apes_leaps_poisson(x, y, mu, maxK)
}
\arguments{
\item{x}{n times p design matrix for logistics regression}

\item{y}{response vector for logistics regression of length n}

\item{mu}{esimated probability of each observation using a baseline model. Of length n.}

\item{maxK}{maximum model size, less than p.}
}
\description{
Perform APES using the leaps package
}
\examples{

set.seed(10)
n = 100
p = 10
maxK = 5
x = matrix(rnorm(n*p), ncol = p)
colnames(x) = paste0("X", 1:p)
y = rpois(n, 5)
mu = glm(y ~ x,family = "poisson")$fitted.value
apesResult = apes_leaps_poisson(x = x, y = y, mu = mu, maxK = maxK)
names(apesResult)
}
