% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apes_logit_mio_combine.R
\name{apes_logit_mio_combine}
\alias{apes_logit_mio_combine}
\title{Bootstrap version of apes_logit}
\usage{
apes_logit_mio_combine(listApes)
}
\description{
Combine apes_logit results
}
\examples{
set.seed(10)
n = 100
p = 10
k = 1:10
beta = c(1, -1, rep(0, p-2))
x = matrix(rnorm(n*p), ncol = p)
colnames(x) = paste0("X", 1:p)
y = rbinom(n = n, size = 1, prob = expit(x \%*\% beta))
Pi = glm.fit(x = x, y = y, family = binomial(link = "logit"))$fitted.values
firstApes = apes_logit(x = x, y = y, Pi = Pi, k = 1, estimator = "mio")
secondApes = apes_logit(x = x, y = y, Pi = Pi, k = 2, estimator = "mio")
listApes = list(firstApes, secondApes)
finalResult = apes_logit_mio_combine(listApes)
print(finalResult)
}
\author{
Kevin Wang
}
