% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apes_mio_poisson.R
\name{apes_mio_poisson}
\alias{apes_mio_poisson}
\title{Perform APES-logistics regression using the bestsubset package}
\usage{
apes_mio_poisson(x, y, mu, krange, time.limit = 100)
}
\arguments{
\item{x}{The n by p design matrix for Poisson regression}

\item{y}{The response vector for Poisson regression of length n}

\item{mu}{Esimated means of each observation using a baseline model. It should be a vector of length n.}

\item{krange}{The range of model size to explore.}
}
\description{
Perform APES-logistics regression using the bestsubset package
}
\examples{

set.seed(10)
n = 100
p = 10
krange = 1:5
x = matrix(rnorm(n*p), ncol = p)
colnames(x) = paste0("X", 1:p)
y = rpois(n, 5)
mu = glm(y ~ x,family = "poisson")$fitted.value
apesResult = apes_mio_poisson(x = x, y = y, mu = mu, krange = krange)
names(apesResult)
}
