% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apes_poisson.R
\name{apes_poisson}
\alias{apes_poisson}
\title{Perform APES-Poisson regression using either the leaps or bestsubset package}
\usage{
apes_poisson(x, y, mu, k, estimator = "leaps", time.limit = 60)
}
\arguments{
\item{x}{The n by p design matrix for Poisson regression}

\item{y}{The response vector for Poisson regression of length n}

\item{mu}{Esimated means of each observation using a baseline model.
It should be a vector of length n.}

\item{k}{Model size to explore. If leaps was selected as the estimator,
then model up to size max(k) will be explored.
If mio was selected as the estimator, then
model sizes specified in k will be explored.}

\item{estimator}{Either "leaps" or "mio", which correspond to
optimisation algorithms available in
the leaps and bestsubset package, respectively.}

\item{time.limit}{The time limit for the maximum time allocated to each
model size model when the "mio" estimator was selected.
It will not affect the speed if leaps}
}
\description{
Perform APES-Poisson regression using either the leaps or bestsubset package
}
\examples{
set.seed(10)
n = 100
p = 10
k = 1:10
beta = c(1, -1, rep(0, p-2))
x = matrix(rnorm(n*p), ncol = p)
colnames(x) = paste0("X", 1:p)
y = rpois(n = n, lambda = exp(x \%*\% beta))
mu = glm.fit(x = x, y = y, family = poisson(link = "log"))$fitted.values
apesLeapsResult = apes_poisson(x = x, y = y, mu = mu, k = k,
                               estimator = "leaps")

apesMioResult = apes_poisson(x = x, y = y, mu = mu, k = k,
                               estimator = "mio", time.limit = 5)

all.equal(apesLeapsResult$apesMleBeta, apesMioResult$apesMleBeta)


}
