% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apes_var_freq.R
\name{apes_var_freq}
\alias{apes_var_freq}
\title{Calculates the frequency that a variable is selected}
\usage{
apes_var_freq(listResult, ic = "BIC")
}
\arguments{
\item{listResult}{a list of APES outputs}

\item{ic}{Either "AIC" or "BIC"}
}
\description{
Calculates the frequency that a variable is selected
}
\examples{
set.seed(10)
n = 100
p = 10
k = 1:10
beta = c(1, -1, rep(0, p-2))
x = matrix(rnorm(n*p), ncol = p)
colnames(x) = paste0("X", 1:p)
y = rpois(n = n, lambda = exp(x \%*\% beta))
mu = glm.fit(x = x, y = y, family = poisson(link = "log"))$fitted.values

listResult = boot_apes_poisson(x = x, y = y, mu = mu, k = k, estimator = "leaps", nBoot = 50)
apes_var_freq(listResult = listResult, ic = "AIC")
apes_var_freq(listResult = listResult, ic = "BIC")
}
\author{
Kevin Wang
}
