% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apesSolver_mio.R
\name{apesSolver_mio}
\alias{apesSolver_mio}
\title{Not to be used by end-users}
\usage{
apesSolver_mio(x, y, k = 1:min(nrow(x), ncol(x)), time.limit)
}
\arguments{
\item{x}{Linearised APES design matrix}

\item{y}{Linearised APES response vector}

\item{k}{The range of model size to explore.}

\item{time.limit}{Time limit in seconds}
}
\description{
Not to be used by end-users
}
