% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_apes_logit.R
\name{boot_apes_logit}
\alias{boot_apes_logit}
\title{Bootstrap version of apes_logit}
\usage{
boot_apes_logit(x, y, Pi, k, estimator = "leaps", time.limit = 60,
  nBoot = 100, nCores = 1)
}
\description{
Bootstrap version of apes_logit
}
\examples{
set.seed(10)
n = 100
p = 10
k = 1:10
beta = c(1, -1, rep(0, p-2))
x = matrix(rnorm(n*p), ncol = p)
colnames(x) = paste0("X", 1:p)
y = rbinom(n = n, size = 1, prob = expit(x \%*\% beta))
Pi = glm.fit(x = x, y = y, family = binomial(link = "logit"))$fitted.values
listResult = boot_apes_logit(x = x, y = y, Pi = Pi, k = k, estimator = "leaps", nBoot = 10)
length(listResult)
}
\author{
Kevin Wang
}
