% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_apes_poisson.R
\name{boot_apes_poisson}
\alias{boot_apes_poisson}
\title{Bootstrap version of apes_logit}
\usage{
boot_apes_poisson(x, y, mu, k, estimator = "leaps", time.limit = 60,
  nBoot = 100, nCores = 1)
}
\description{
Bootstrap version of apes_logit
}
\examples{
set.seed(10)
n = 100
p = 10
k = 1:10
beta = c(1, -1, rep(0, p-2))
x = matrix(rnorm(n*p), ncol = p)
colnames(x) = paste0("X", 1:p)
y = rpois(n = n, lambda = exp(x \%*\% beta))
mu = glm.fit(x = x, y = y, family = poisson(link = "log"))$fitted.values
listResult = boot_apes_poisson(x = x, y = y, mu = mu, k = k, estimator = "leaps", nBoot = 10)
length(listResult)
}
\author{
Kevin Wang
}
