% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icOptimal.R
\name{icOptimal}
\alias{icOptimal}
\title{Not to be used by end-users}
\usage{
icOptimal(ic, symbol)
}
\arguments{
\item{ic}{numeric vector, information criterion}

\item{symbol}{Labelling minimum model with a symbol}
}
\description{
Not to be used by end-users
}
