% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maWeightsPlot.R
\name{maWeightsPlot}
\alias{maWeightsPlot}
\title{MA weights visualisation}
\usage{
maWeightsPlot(listResult, type = "aic")
}
\arguments{
\item{listResult}{a list of APES outputs}

\item{type}{either "aic" (default) or "bic"}
}
\description{
MA weights visualisation
}
\examples{
set.seed(10)
n = 100
p = 10
k = 1:10
beta = c(1, -1, rep(0, p-2))
x = matrix(rnorm(n*p), ncol = p)
colnames(x) = paste0("X", 1:p)
y = rpois(n = n, lambda = exp(x \%*\% beta))
mu = glm.fit(x = x, y = y, family = poisson(link = "log"))$fitted.values

listResult = boot_apes_poisson(x = x, y = y, mu = mu, k = k, estimator = "leaps", nBoot = 50)
maWeightsPlot(listResult, type = "aic")
maWeightsPlot(listResult, type = "bic")
}
\author{
Kevin Wang
}
