% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minIcMatrix.R
\name{minIcMatrix}
\alias{minIcMatrix}
\title{Not to be used by end-users}
\usage{
minIcMatrix(ic, mat)
}
\arguments{
\item{ic}{numeric vector, information criterion}

\item{mat}{matrix}
}
\description{
Not to be used by end-users
}
