% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mleModelToBeta.R
\name{mleModelToBeta}
\alias{mleModelToBeta}
\title{We take the MLE model fitted above and rearrange the columns and remove the NA's in our vector}
\usage{
mleModelToBeta(mleModels, variables)
}
\arguments{
\item{mleModels}{mleModels}

\item{variables}{variable names}
}
\description{
We take the MLE model fitted above and rearrange the columns and remove the NA's in our vector
}
