% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot_apes.R
\name{mplot_boot}
\alias{mplot_boot}
\title{Plotting APES models uing mplot (experimental)}
\usage{
mplot_boot(boot_apes, model, interactive = FALSE)
}
\arguments{
\item{boot_apes}{Bootstrapped APES result}

\item{model}{The full model}

\item{interactive}{Default to FALSE.
If TRUE, then googleVis will be used to make an interactive plot.}
}
\description{
Plotting APES models uing mplot (experimental)
}
\examples{
set.seed(10)
n = 100
p = 10
k = 1:10
beta = c(1, -1, rep(0, p-2))
x = matrix(rnorm(n*p), ncol = p)
colnames(x) = paste0("X", 1:p)

y = rbinom(n = n, size = 1, prob = expit(x \%*\% beta))
data = data.frame(y, x)
model = glm(y ~ ., data = data, family = "binomial")
boot_apes = apes(model = model, n_boot = 20)
library(mplot)
mplot_boot(boot_apes = boot_apes, model = model)
mplot_boot(boot_apes = boot_apes, model = model, interactive = TRUE)
}
