% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.apes.R
\name{plot.apes}
\alias{plot.apes}
\title{Generic plotting function for class \code{apes}}
\usage{
\method{plot}{apes}(x, type = "path", order = c("BIC", "AIC", "median"), ...)
}
\arguments{
\item{x}{An object of class \code{apes}}

\item{type}{Type of plot: (only supporting "ic" at the moment)
\itemize{
\item{"path"}: Information criterion vs model size.
Either "AIC" or "BIC" can be shown using the order argument.
}}

\item{order}{Either "AIC", "BIC". If type is selected to be "vi_tile", then also takes the value "median".}

\item{...}{additional parameters (not currently used)}
}
\value{
A ggplot output corresponding to the select plotting type.
}
\description{
Generic plotting function for class \code{apes}
}
\examples{
set.seed(10)
n = 100
p = 10
k = 1:10
beta = c(1, -1, rep(0, p-2))
x = matrix(rnorm(n*p), ncol = p)
colnames(x) = paste0("X", 1:p)
y = rbinom(n = n, size = 1, prob = expit(x \%*\% beta))
data = data.frame(y, x)
model = glm(y ~ ., data = data, family = "binomial")

apes_result = apes(model = model)
plot(apes_result)
}
