% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_apes_ic.R
\name{plot_apes_ic}
\alias{plot_apes_ic}
\title{AIC or BIC path plot}
\usage{
plot_apes_ic(list_result, type = "BIC")
}
\arguments{
\item{list_result}{a list of APES outputs}

\item{type}{Either "AIC" or "BIC"}
}
\description{
AIC or BIC path plot
}
\examples{
set.seed(10)
n = 100
p = 10
k = 1:10
beta = c(1, -1, rep(0, p-2))
x = matrix(rnorm(n*p), ncol = p)
colnames(x) = paste0("X", 1:p)
y = rbinom(n = n, size = 1, prob = expit(x \%*\% beta))
data = data.frame(y, x)
model = glm(y ~ ., data = data, family = "binomial")

list_result = apes(model = model, n_boot = 20)

plot_apes_ic(list_result = list_result)
}
\author{
Kevin Wang
}
