% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ma_coef.R
\name{plot_ma_coef}
\alias{plot_ma_coef}
\title{Average of model average plot}
\usage{
plot_ma_coef(list_result, type = "AIC")
}
\arguments{
\item{list_result}{a list of APES outputs}

\item{type}{either "AIC" (default) or "BIC"}
}
\description{
This function is suitable for a list of bootstrap APES outputs.
From each bootstrap run, APES stores coefficient values averaged across all models considered.
As we have multiple bootstrapped APES output, we can cummulatively average
these model averaged coefficient values across all bootstrap runs.
On the final plot, we should be able to see variables of non-zero coefficients show up distinctly away from zero.
}
\details{
Visualise model averaging results
}
\examples{
set.seed(10)
n = 100
p = 10
k = 1:10
beta = c(1, -1, rep(0, p-2))
x = matrix(rnorm(n*p), ncol = p)
colnames(x) = paste0("X", 1:p)
y = rbinom(n = n, size = 1, prob = expit(x \%*\% beta))
data = data.frame(y, x)
model = glm(y ~ ., data = data, family = "binomial")

list_result = apes(model = model, n_boot = 20)

plot_ma_coef(list_result = list_result)
}
\author{
Kevin Wang
}
