% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vi.R
\name{plot_vi}
\alias{plot_vi}
\title{Variable Inclusion Plot}
\usage{
plot_vi(listResult)
}
\arguments{
\item{listResult}{a list of APES outputs}
}
\value{
bootVarPlotdf a tibble (data.frame) with all the necessary values to plot a variable inclusion plot

viPlot a ggplot
}
\description{
This function is suitable for a list of bootstrap APES outputs.
From each bootstrap run, APES stores log-likelihood for every model it considered.
In this function, we then consider general information criterion (GIC) of the form
-2*logLike + penalty * modeSize.
For each penalty and each bootstrap run, we apply this GIC to find a model of the optimal fit, and then
look at which variables are selected in that model.
The frequency of a variable selected across different penalties are then avaraged across all bootstrap runs.
}
\details{
Variable inclusion plot
}
\examples{
set.seed(10)
n = 100
p = 10
k = 1:10
beta = c(1, -1, rep(0, p-2))
x = matrix(rnorm(n*p), ncol = p)
colnames(x) = paste0("X", 1:p)
y = rpois(n = n, lambda = exp(x \%*\% beta))
mu = glm.fit(x = x, y = y, family = poisson(link = "log"))$fitted.values

listResult = boot_apes_poisson(x = x, y = y, mu = mu, k = k, estimator = "leaps", nBoot = 20)
viPlotResult = plot_vi(listResult)
viPlotResult$viPlot
}
\author{
Kevin Wang
}
