% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vi.R
\name{plot_vi_boot_apes}
\alias{plot_vi_boot_apes}
\title{Variable Inclusion Plot}
\usage{
plot_vi_boot_apes(list_result)
}
\arguments{
\item{list_result}{a list of APES outputs}
}
\value{
boot_vars_plotdf a tibble (data.frame) with all the necessary values to plot a variable inclusion plot

vi_plot a ggplot
}
\description{
This function is suitable for a list of bootstrap APES outputs.
From each bootstrap run, APES stores log-likelihood for every model it considered.
In this function, we then consider general information criterion (GIC) of the form
-2*logLike + penalty * modeSize.
For each penalty and each bootstrap run, we apply this GIC to find a model of the optimal fit, and then
look at which variables are selected in that model.
The frequency of a variable selected across different penalties are then avaraged across all bootstrap runs.
}
\details{
Variable inclusion plot
}
\examples{
set.seed(10)
n = 100
p = 10
k = 1:10
beta = c(1, -1, rep(0, p-2))
x = matrix(rnorm(n*p), ncol = p)
colnames(x) = paste0("X", 1:p)
y = rbinom(n = n, size = 1, prob = expit(x \%*\% beta))
data = data.frame(y, x)
model = glm(y ~ ., data = data, family = "binomial")

list_result = apes(model = model, n_boot = 50)

plot_vi_boot_apes(list_result = list_result)
}
\author{
Kevin Wang
}
