% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vi_tile.R
\name{plot_vi_tile}
\alias{plot_vi_tile}
\title{Variable Inclusion Plot in tile format}
\usage{
plot_vi_tile(listResult)
}
\arguments{
\item{listResult}{a list of APES outputs}
}
\value{
apesMleBetaBinaryPlotdf a tibble (data.frame) with all the necessary values to plot a variable inclusion plot

variableTilePlot a ggplot with continuous colouring

variableTilePlot_category a ggplot with discrete colouring
}
\description{
This function displays the same information as plot_vi, but in a tile plot format.
}
\details{
Variable inclusion plot in tile format
}
\examples{
set.seed(10)
n = 100
p = 10
k = 1:10
beta = c(1, -1, rep(0, p-2))
x = matrix(rnorm(n*p), ncol = p)
colnames(x) = paste0("X", 1:p)
y = rpois(n = n, lambda = exp(x \%*\% beta))
mu = glm.fit(x = x, y = y, family = poisson(link = "log"))$fitted.values

listResult = boot_apes_poisson(
x = x, y = y, mu = mu, k = k,
estimator = "leaps", nBoot = 20)
viTileResult = plot_vi_tile(listResult)
viTileResult$variableTilePlot
viTileResult$variableTilePlot_category
}
\author{
Kevin Wang
}
