% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vi_tile.R
\name{plot_vi_tile_boot_apes}
\alias{plot_vi_tile_boot_apes}
\title{Variable Inclusion Plot in tile format}
\usage{
plot_vi_tile_boot_apes(list_result, order = "median")
}
\arguments{
\item{list_result}{a list of APES outputs}

\item{order}{The ordering of variables. Either "median", "AIC" or "BIC"}
}
\value{
apesMleBetaBinaryPlotdf a tibble (data.frame) with all the necessary values to plot a variable inclusion plot

variableTilePlot a ggplot with continuous colouring

variableTilePlot_category a ggplot with discrete colouring
}
\description{
This function displays the same information as plot_vi, but in a tile plot format.
}
\details{
Variable inclusion plot in tile format
}
\examples{
set.seed(10)
n = 100
p = 10
k = 1:10
beta = c(1, -1, rep(0, p-2))
x = matrix(rnorm(n*p), ncol = p)
colnames(x) = paste0("X", 1:p)
y = rbinom(n = n, size = 1, prob = expit(x \%*\% beta))
data = data.frame(y, x)
model = glm(y ~ ., data = data, family = "binomial")

list_result = apes(model = model, n_boot = 20)

plot_vi_tile_boot_apes(list_result = list_result)
}
\author{
Kevin Wang
}
