% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit-expit.R
\name{logit}
\alias{logit}
\alias{expit}
\title{The logit function}
\usage{
logit(x)

expit(x)
}
\arguments{
\item{x}{numeric}
}
\description{
The logit function

The expit function
}
\examples{
curve(logit, from = 0.1, to = 0.9)
curve(expit, from = -5, to = 5)
}
