% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apes_logit.R
\name{apes_logit}
\alias{apes_logit}
\title{Perform APES-logistic regression using either the leaps or bestsubset package}
\usage{
apes_logit(x, y, Pi, k, estimator = "leaps", time.limit = 60)
}
\arguments{
\item{x}{The n by p design matrix for logistic regression}

\item{y}{The response vector for logistic regression of length n}

\item{Pi}{Esimated probabilities of each observation using a baseline model. It should be a vector of length n.}

\item{k}{Model size to explore. If leaps was selected as the estimator, then model up to size max(k) will be explored. If mio was selected as the estimator, then model sizes specified in k will be explored.}

\item{estimator}{Either "leaps" or "mio", which correspond to optimisation algorithms available in the leaps and bestsubset package, respectively.}

\item{time.limit}{The time limit for the maximum time allocated to each model size model when the "mio" estimator was selected. It will not affect the speed if leaps}
}
\value{
A list of results:
\itemize{
\item{apesModelDf:} A data.frame (tibble) containing the optimal model of each size and associated statistics
\item{apesMleBeta:} A matrix of MLE for the most optimal model of each model size.
\item{apesMleBetaBinary:} A data.frame (tibble), which has a column indicating if
a variable has been selected into the most optimal model of each size.
It is a binarised, linear version of the previous \code{apesMleBeta} output.
\item{apesTimeDiff:} A difftime object of the computational time.
\item{selectedModelBeta:} A matrix containing the MLE correspond to the AIC and BIC selected model. A subset of the \code{apesMleBeta} output.
\item{aicWeights:} A matrix of weights corresponding to the AIC_j/sum(AIC_j), where AIC_j denotes the most optimal model of the size j in the \code{apesModelDf} output.
\item{bicWeights:} A matrix of weights corresponding to the BIC_j/sum(BIC_j), where BIC_j denotes the most optimal model of the size j in the \code{apesModelDf} output.
\item{modelAvgBeta}: A matrix of AIC and BIC weighted coefficients. It is identical to the matrix product of apesMleBeta and aicWeights (or bicWeights).
\item{responseTibble}: A data.frame (tibble) of estimates for each observations, e.g. initial fitted response, final fitted response.
}
}
\description{
Perform APES-logistic regression using either the leaps or bestsubset package
}
\examples{
set.seed(10)
n = 100
p = 10
k = 1:10
beta = c(1, -1, rep(0, p-2))
x = matrix(rnorm(n*p), ncol = p)
colnames(x) = paste0("X", 1:p)
y = rbinom(n = n, size = 1, prob = expit(x \%*\% beta))
Pi = glm.fit(x = x, y = y, family = binomial(link = "logit"))$fitted.values
apesLeapsResult = apes_logit(x = x, y = y, Pi = Pi, k = k,
                               estimator = "leaps")

apesMioResult = apes_logit(x = x, y = y, Pi = Pi, k = k,
                               estimator = "mio", time.limit = 5)

all.equal(apesLeapsResult$apesMleBeta, apesMioResult$apesMleBeta)


}
