% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apes_mio_logit.R
\name{apes_mio_logit}
\alias{apes_mio_logit}
\title{Perform APES using the bestsubset package}
\usage{
apes_mio_logit(x, y, Pi, krange, time.limit = 100)
}
\arguments{
\item{x}{n times p design matrix for logistics regression}

\item{y}{response vector for logistics regression of length n}

\item{Pi}{esimated probability of each observation using a baseline model. Of length n.}

\item{krange}{range of k to explore}
}
\description{
Perform APES using the bestsubset package
}
\examples{
set.seed(10)
n = 100
p = 10
krange = 1:5
x = matrix(rnorm(n*p), ncol = p)
colnames(x) = paste0("X", 1:p)
y = round(runif(n, 0,1))
Pi = runif(n, 0, 1)
apesResult = apes_mio(x = x, y = y, Pi = Pi, krange = krange)
names(apesResult)
}
