% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apes_var_freq.R
\name{apes_var_freq}
\alias{apes_var_freq}
\title{Calculates the frequency that a variable is selected}
\usage{
apes_var_freq(list_result, ic = "BIC")
}
\arguments{
\item{list_result}{a list of APES outputs}

\item{ic}{Either "AIC" or "BIC"}
}
\description{
Calculates the frequency that a variable is selected
}
\examples{
set.seed(10)
n = 100
p = 10
k = 1:10
beta = c(1, -1, rep(0, p-2))
x = matrix(rnorm(n*p), ncol = p)
colnames(x) = paste0("X", 1:p)
y = rbinom(n = n, size = 1, prob = expit(x \%*\% beta))
data = data.frame(y, x)
model = glm(y ~ ., data = data, family = "binomial")

list_result = apes(model = model, n_boot = 20)

apes_var_freq(list_result = list_result, ic = "AIC")
apes_var_freq(list_result = list_result, ic = "BIC")
}
\author{
Kevin Wang
}
