% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_apes_logit.R
\name{boot_apes_logit}
\alias{boot_apes_logit}
\title{Bootstrap version of apes_logit}
\usage{
boot_apes_logit(x, y, Pi, k, estimator = "leaps", time.limit = 60,
  nBoot = 100, nCores = 1)
}
\arguments{
\item{x}{The n by p design matrix for logistic regression}

\item{y}{The response vector for logistic regression of length n}

\item{Pi}{Esimated probabilities of each observation using a baseline model. It should be a vector of length n.}

\item{k}{Model size to explore. If leaps was selected as the estimator,
then model up to size max(k) will be explored.
If mio was selected as the estimator, then
model sizes specified in k will be explored.}

\item{estimator}{Either "leaps" or "mio", which correspond to
optimisation algorithms available in the leaps and bestsubset package, respectively.}

\item{time.limit}{The time limit for the maximum time allocated
to each model size model when the "mio" estimator was selected.
It will not affect the speed if leaps}

\item{nBoot}{Number of bootstrap runs}

\item{nCores}{Number of cores, only supporting parallel::mclapply now}
}
\description{
Bootstrap version of apes_logit
}
\examples{
set.seed(10)
n = 100
p = 10
k = 1:10
beta = c(1, -1, rep(0, p-2))
x = matrix(rnorm(n*p), ncol = p)
colnames(x) = paste0("X", 1:p)
y = rbinom(n = n, size = 1, prob = expit(x \%*\% beta))
Pi = glm.fit(x = x, y = y, family = binomial(link = "logit"))$fitted.values
listResult = boot_apes_logit(
x = x, y = y, Pi = Pi, k = k,
estimator = "leaps", nBoot = 10, nCores = 1)
length(listResult)
}
\author{
Kevin Wang
}
