% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_apes_poisson.R
\name{boot_apes_poisson}
\alias{boot_apes_poisson}
\title{Bootstrap version of apes_logit}
\usage{
boot_apes_poisson(x, y, mu, k, estimator = "leaps", time.limit = 60,
  nBoot = 100, workers = 1)
}
\arguments{
\item{x}{The n by p design matrix for Poisson regression}

\item{y}{The response vector for Poisson regression of length n}

\item{mu}{Esimated means of each observation using a baseline model.
It should be a vector of length n.}

\item{k}{Model size to explore. If leaps was selected as the estimator,
then model up to size max(k) will be explored.
If mio was selected as the estimator, then
model sizes specified in k will be explored.}

\item{estimator}{Either "leaps" or "mio", which correspond to
optimisation algorithms available in
the leaps and bestsubset package, respectively.}

\item{time.limit}{The time limit for the maximum time allocated to each
model size model when the "mio" estimator was selected.
It will not affect the speed if leaps}

\item{nBoot}{Number of bootstrap runs}

\item{workers}{Number of cores, using multicores from the furr package}
}
\description{
Bootstrap version of apes_logit
}
\examples{
set.seed(10)
n = 100
p = 10
k = 1:10
beta = c(1, -1, rep(0, p-2))
x = matrix(rnorm(n*p), ncol = p)
colnames(x) = paste0("X", 1:p)
y = rpois(n = n, lambda = exp(x \%*\% beta))
mu = glm.fit(x = x, y = y, family = poisson(link = "log"))$fitted.values
listResult = boot_apes_poisson(
x = x, y = y, mu = mu, k = k,
estimator = "leaps", nBoot = 10, workers = 1)
length(listResult)
}
\author{
Kevin Wang
}
