% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mleModelToBeta.R
\name{mleModelToBeta}
\alias{mleModelToBeta}
\title{Convert MLE models to beta coefficients}
\usage{
mleModelToBeta(mleModels, variables)
}
\arguments{
\item{mleModels}{mleModels}

\item{variables}{variable names}
}
\description{
Convert MLE models to beta coefficients
}
